In-memory engine: strip off read related code from engine.rs to a separte file (#16792)
ref tikv/tikv#16141

strip off read related code from engine.rs to a separte file
